\name{CXCR}
\alias{CXCR}
\alias{cxcr4}
\alias{cxcr7}
\alias{shh}

\title{
  Clinical relevance of CXCL12/CXCR4 and Hedgehog
}
\encoding{UTF-8}
\description{
  Determine the clinical relevance of CXCL12/CXCR4 and 
  Hedgehog interaction in pancreatic cancer.
}
\usage{
  data(CXCR)
}

\format{
  
  Three small data set: \code{cxcr4}, \code{cxcr7},
  and \code{shh}.
}

\references{

 Sanjeev K. Srivastava , Arun Bhardwaj , Seema Singh , Sumit Arora ,
 Bin Wang , William E. Grizzle and Ajay P. Singh (2011) "MicroRNA-150
 directly targets MUC4 and suppresses growth and malignant behavior of
 pancreatic cancer cells", Carcinogenesis,
 doi:10.1093/carcin/bgr223. (PMID 21983127).

 Arora S., Bhardwaj A., Srivastava S.K., Singh S., McClellan S., Wang
  B., Singh A.P. (2011) "Honokiol arrests cell cycle, induces apoptosis
  and potentiates the cytotoxic effects of gemcitabine in human
  pancreatic cancer cells", PLoS One, 6 (6), e21573.

}

\examples{
  data(CXCR)
  shh
}
\keyword{data}
