\name{illuRNA}
\docType{data}
\alias{illuRNA}
\alias{rna4}
\alias{print.ilRNA}

\usage{
  data(illuRNA)
}

\title{A sample data of two-color microarray}
\description{
 CGH microarray.
}

\format{

  A data frame with 44495 rows and 19 columns (variables).

  Class type: \code{ilRNA}, data name is \code{rna4}.
}

\references{

 Wang, B., Zhang, S-G., Wang, X-F., Tan, M. and Xi, Y. (2012) "Testing
 for differentially-expressed microRNAs with errors-in-variables
 nonparametric regression", PLoS ONE 7(5):
 e37537. doi:10.1371/journal.pone.0037537.

 Wang, B., Wang, X-F. and Xi, Y. (2011) "Normalizing bead-based
 microRNA expression data: a measurement error model-based approach",
 Bioinformatics, 27(11), 1506-1512.

 Wang, B., Howell, P., Bruheim, S. Ju, J, Owen, L.B., Fodstad, O. and
 Xi, Y. (2011). "Systematic Evaluation of Three microRNA Profiling
 Platforms: Microarray, Beads Array, and Quantitative Real-Time PCR
 Array", PLoS ONE 6(2): e17167.

 Wang, B. and Wertelecki, W. (2012) "Density Estimation for Data With
 Rounding Errors", Computational Statistics and Data Analysis, (in
 press). doi: 10.1016/j.csda.2012.02.016. link

 Wang, B., Wang, X-F. and Xi, Y. (2011) "Normalizing bead-based
 microRNA expression data: a measurement error model-based approach",
 Bioinformatics, 27(11), 1506-1512.

 Wang, X-F. and Wang, B. (2011) "Deconvolution Estimation in
 Measurement Error Models: The R Package decon", Journal of Statistical
 Software, 39(10), 1-24.

}

\examples{
  data(illuRNA)
}

\keyword{datasets}
