\name{ofc}
\docType{data}
\alias{ofc}
\usage{data(ofc)}
\title{occipitofrontal head circumference data}
\description{
OFC data for singleton live births with gestational age at least 38 weeks.
}
\format{A data frame with 2019 observations on 4 variables.
  \tabular{lll}{
    \code{Year}  \tab numeric  \tab 2006 -- 2009 \cr
    \code{Sex}   \tab character  \tab 'male' or 'female' \cr
    \code{Gestation}  \tab numeric  \tab Gestational age (in weeks). \cr
    \code{Head}   \tab numeric  \tab head size. \cr
  }
}
\references{

 Wang, B. and Wertelecki, W. (2012) "Density Estimation for Data With
 Rounding Errors", Computational Statistics and Data Analysis, (in
 press). doi: 10.1016/j.csda.2012.02.016. link

}

\keyword{datasets}
